Refactor the sighash module
Done in preparation for splitting the `EcdsaSighashType` up into legacy
and segwit v0.

Refactor the `sighash` module splitting the cache out into its own
module and the sighash types into their own modules (currently `taproot`
and `ecdsa`).

Make the new modules private and export types from the `sighash/mod.rs`
file so other parts of the codebase remain unchanged.

Refactor only, no code or logic changes outside of adding modules and
exporting types.
